requirements.txt
fastapi
streamlit
easyocr
Pillow
pandas
transformers
sentencepiece
tensorflow
scikit-learn